+++ test/testscripts/subplot_grid.jl
@@ -3,9 +3,9 @@ using RDatasets, DataFrames, Gadfly
 set_default_plot_size(10inch, 10inch)
 
 barley = dataset("lattice", "barley")
-levels!(barley[:Year], ["1931", "1932"])
+levels!(barley.Year, ["1931", "1932"])
 
-idx = [startswith(x,"No.") for x in barley[:Variety]]
+idx = [startswith(x,"No.") for x in barley.Variety]
 plot(barley[idx,:],
      xgroup="Variety", ygroup="Site", x="Year", y="Yield",
      Geom.subplot_grid(Geom.line, Geom.point))